# cache SSL client certs