benchmark                                       old ns/op     new ns/op     delta
    BenchmarkSetupSSLClientCertificatesParse-8      379742        77.8          -99.98%
    BenchmarkSetupSSLClientCertificatesNoFile-8     2592          59.7          -97.70%
    
    benchmark                                       old allocs     new allocs     delta
    BenchmarkSetupSSLClientCertificatesParse-8      2702           0              -100.00%
    BenchmarkSetupSSLClientCertificatesNoFile-8     15             0              -100.00%
    
    benchmark                                       old bytes     new bytes     delta
    BenchmarkSetupSSLClientCertificatesParse-8      206032        0             -100.00%
    BenchmarkSetupSSLClientCertificatesNoFile-8     680           0             -100.00%