@kr 2015-10-05
Not sure what's going on with the builds from Go tip. This patch doesn't touch anything related to the tests that are failing.