@kr 2015-10-05
The travis failures appear to be unrelated to this change. See #398.