@johto 2015-10-06
Gah.  I'd already forgotten that there's no way to associate anything with an *sql.DB instance with the way database/sql has been designed.

I'm really not sure this is a good idea, at least as currently written.  Any application which previously allowed you to reload certificates by letting go of its *sql.DB handle and creating a new one will suddenly not be able to do that at all.  That seems like a big issue.