@johto 2015-10-09
> Gah. I'd already forgotten that there's no way to associate anything with an *sql.DB instance with the way database/sql has been designed.

I wonder if a better way forward would be to suggest a fix for this problem first.  The drivers could implement an optional interface which would be called on sql.Open(), and instead of storing the shared "driver" object, DB.driver would be set to a new object returned by this new suggested method.  A driver factory of a kind, if you will.  The interface of Open already supports that since you can it can return an error.