@kr 2016-02-08
The cert and key are global data. I don't see what they have to do with a *sql.DB. The same cert can and should be used for more than one *sql.DB instance.

> reload certificates by letting go of its *sql.DB handle

This seems like a peculiar way to load new global certificate data. This package certainly never worked that way before. As a user, I'd be surprised to see such behavior. Actually, I'd probably never need to do it in the first place, but if I did, I'd expect to call a function like pq.ReloadCerts, or, preferably, to have it match the current behavior: to load the current cert data for each new connection.

For my purposes, I don't really need this patch any more, so I'll just abandon it here. But anyone else feel free to pick it up and take it across the line.